+++ src/resources/parsers/scene1.ts
@@ -317,10 +317,11 @@ function loadPoints(scene, offset) {
         const point = {
             sceneIndex: scene.index,
             index: i,
+            // normalise position
             pos: [
-                ((0x8000 - data.getInt16(offset + 4, true)) + 512) * WORLD_SCALE,
+                (((0x8000 - data.getInt16(offset + 4, true)) + 512) - 256) * WORLD_SCALE,
                 data.getInt16(offset + 2, true) * WORLD_SCALE,
-                data.getInt16(offset, true) * WORLD_SCALE
+                (data.getInt16(offset, true) + 256) * WORLD_SCALE
             ]
         };
         offset += 6;