+++ CHANGELOG.md
@@ -15,6 +15,7 @@ _2023_
 
 - LBA1 Climbing Ladders
 - Extra bonus speed
+- Behavioiur Menu Icons CSS
 
 
 ## [v0.6.0](https://github.com/LBALab/lba2remake/compare/v0.5.0...v0.5.1) Big update!